Fix loading order for javax.cache.CacheManager on App Engine
I was running an drawio fork on app engine and Google Drive integration suddenly started failing with IncompatibleClassChangeError errors. Found https://stackoverflow.com/q/46582783 which fixed it.

We might not 100% need this. This is mostly a heads up if anyone else runs into similar issues.
I'm also not sure if this is easily reproducible on the public App Engine. Following up with GAE to see if this is a bigger issue.